Move entities to normal namespace
