+++ StressCLI.Tests/ResultWriterTests.cs
@@ -1,6 +1,6 @@
 ﻿using StressCLI.src.Helpers;
-using StressCLI.src.TestCore;
-using StressCLI.src.TestCore.ResultSetter;
+using StressCLI.src.Entities;
+using StressCLI.src.Entities.ResultSetter;
 using System;
 using System.Collections.Generic;
 using System.IO;
