+++ StressCLI/src/TestCore/Interfaces/IConfigParser.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Text;
 
-namespace StressCLI.src.TestCore.Interfaces
+namespace StressCLI.src.Entities.Interfaces
 {
     public interface IConfigParser
     {
